requirements.txt
chromadb==0.6.1
langchain_community==0.3.0
langchain==0.3.0
requests==2.32.3
pypdf==5.0.0
PyPDF3==1.0.6
python-dotenv==1.0.1
streamlit==1.38.0
langchain-openai==0.2.9
langchain-ollama==0.2.2
docling==2.8.3
python-pptx==1.0.2
python-docx==1.1.2
